generatePresentationMultiple

Generates a presentation from a Strategus result

generatePresentationMultiple(
  server,
  username,
  password,
  dbms,
  resultsSchema = NULL,
  targetId,
  subsetId = NULL,
  includeIndication = TRUE,
  outcomeIds,
  comparatorIds,
  covariateIds = NULL,
  friendlyNames = list(targetName = "target cohort", comparatorNames =
    c("comparator cohort 1", "comparator cohort 2"), indicationName =
    "indication cohort", outcomeNames = c("outcome name 1", "outcome name 2")),
  details = list(studyPeriod = "All Time", restrictions = "Age - None"),
  title = "ASSURE 001 ...",
  lead = "add name",
  date = Sys.Date(),
  backgroundText = "",
  evaluationText = "",
  outputLocation = getwd(),
  outputName = paste0("presentation_", gsub(":", "_", gsub(" ", "_",
    as.character(date()))), ".html"),
  intermediateDir = tempdir()
)

Arguments

server

The server containing the result database

username

The username for an account that can access the result database

password

The password for an account that can access the result database

dbms

The dbms used to access the result database

resultsSchema

The result database schema

targetId

The cohort definition id for the target cohort

subsetId

Optional a subset ID for the cohort method/characterization and prediction results

includeIndication

Whether an indication was used in this study

outcomeIds

The cohort definition id for the outcome

comparatorIds

The cohort method comparator cohort id

covariateIds

A vector of covariateIds to include in the characterization

friendlyNames

a data.frame with friendly name conversions

details

a list with the studyPeriod and restrictions

title

A title for the presentation

lead

The name of the presentor

date

The date of the presentation

backgroundText

a character with any background text

evaluationText

a list of bullet points for the evaluation

outputLocation

The file location and name to save the protocol

outputName

The name of the html protocol that is created

intermediateDir

The work directory for quarto

Value

An named R list with the elements 'standard' and 'source'

Details

Specify the connection details to the result database and the schema name to generate a presentation.