A function to extract case series characterization results

A function to extract case series characterization results

Usage

getBinaryCaseSeries(
  connectionHandler,
  schema,
  cTablePrefix = "c_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  targetId = NULL,
  outcomeId = NULL,
  databaseIds = NULL,
  riskWindowStart = NULL,
  riskWindowEnd = NULL,
  startAnchor = NULL,
  endAnchor = NULL,
  conceptIds = NULL,
  minVal = NULL
)

Arguments

connectionHandler

A connection handler that connects to the database and extracts sql queries. Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.

schema

The result database schema (e.g., 'main' for sqlite)

cTablePrefix

The prefix used for the characterization results tables

cgTablePrefix

The prefix used for the cohort generator results tables

databaseTable

The name of the table with the database details (default 'database_meta_data')

targetId

An integer corresponding to the target cohort ID

outcomeId

Am integer corresponding to the outcome cohort ID

databaseIds

(optional) One or more unique identifiers for the databases

riskWindowStart

(optional) A riskWindowStart to restrict to

riskWindowEnd

(optional) A riskWindowEnd to restrict to

startAnchor

(optional) A startAnchor to restrict to

endAnchor

(optional) An endAnchor to restrict to

conceptIds

(optional) An conceptIds to restrict to

minVal

(optional) the minimum averageVal to extract

Value

A data.frame with the characterization case series results

Details

Specify the connectionHandler, the schema and the target/outcome cohort IDs

See also

Examples

conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)
#> Connecting using SQLite driver

cs <- getBinaryCaseSeries(
  connectionHandler = connectionHandler, 
  schema = 'main',
  targetId = 1, 
  outcomeId = 3
)