A function to extract the targets found in incidence

A function to extract the targets found in incidence

Usage

getIncidenceTargets(
  connectionHandler,
  schema,
  ciTablePrefix = "ci_",
  cgTablePrefix = "cg_"
)

Arguments

connectionHandler

A connection handler that connects to the database and extracts sql queries. Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.

schema

The result database schema (e.g., 'main' for sqlite)

ciTablePrefix

The prefix used for the cohort incidence results tables

cgTablePrefix

The prefix used for the cohort generator results tables

Value

A data.frame with the incidence target cohort ids and names

Details

Specify the connectionHandler, the schema and the prefixes

See also

Examples

conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)
#> Connecting using SQLite driver

cohorts <- getIncidenceTargets(
  connectionHandler = connectionHandler, 
  schema = 'main'
)