Extract model interception (for logistic regression)

This function extracts the interception value

Usage

getPredictionIntercept(
  connectionHandler,
  schema,
  plpTablePrefix = "plp_",
  modelDesignId = NULL,
  databaseId = NULL
)

Arguments

connectionHandler

A connection handler that connects to the database and extracts sql queries. Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.

schema

The result database schema (e.g., 'main' for sqlite)

plpTablePrefix

The prefix used for the patient level prediction results tables

modelDesignId

The identifier for a model design to restrict to

databaseId

The identifier for the development database to restrict to

Value

Returns a single value corresponding to the model intercept or NULL if not a logistic regression model

Details

Specify the connectionHandler, the resultDatabaseSettings, the modelDesignId and the databaseId

See also

Examples

conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)
#> Connecting using SQLite driver

intercepts <- getPredictionIntercept(
  connectionHandler = connectionHandler, 
  schema = 'main'
)
#> Warning: Please enter a modelDesignId and databaseId