A function to extract the outcomes found in self controlled case series

A function to extract the outcomes found in self controlled case series

Usage

getSccsOutcomes(
  connectionHandler,
  schema,
  sccsTablePrefix = "sccs_",
  cgTablePrefix = "cg_",
  targetId = NULL
)

Arguments

connectionHandler

A connection handler that connects to the database and extracts sql queries. Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.

schema

The result database schema (e.g., 'main' for sqlite)

sccsTablePrefix

The prefix used for the cohort generator results tables

cgTablePrefix

The prefix used for the cohort generator results tables

targetId

An integer corresponding to the target cohort ID

Value

A data.frame with the self controlled case series outcome ids and names.

Details

Specify the connectionHandler, the schema and the prefixes

See also

Examples

conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)
#> Connecting using SQLite driver

outcomes <- getSccsOutcomes(
  connectionHandler = connectionHandler, 
  schema = 'main'
)