View the Incidence Rates

Creates a table with the incidence rates and optionally demographics

Usage

viewIncidenceRate(
  incidenceData,
  ageData = NULL,
  genderData = NULL,
  stratification = "overall",
  maxAgeSampleSize = 5000
)

Arguments

incidenceData

The data extracted using 'getIncidenceRates'

ageData

The data extracted using 'getBinaryTargetBaseline' with analysisIds = 3

genderData

The data extracted using 'getBinaryTargetBaseline' with analysisIds = 1

stratification

Pick either overall/age/sex/year to specify whether to view the overall rates or stratified by age/sex/year

maxAgeSampleSize

When creating the age distributions this is the max age vector length to create

Value

Returns a gt table that displays the incidence rates

Details

Input the incidence rate data (and optionally demographic data)

See also

Examples

conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)
#> Connecting using SQLite driver
schema <- 'main'

incidenceData <- getIncidenceRates(
  connectionHandler = connectionHandler , 
  schema = schema
  )
  
  # incidence data does not have rate values to imputing them
  incidenceData$incidenceRateP100py <- 1 +
    sample(c(-1,1),replace = TRUE)*runif(nrow(incidenceData))
  incidenceData$incidenceProportionP100p <- 0.5 +
    sample(c(-1,1),replace = TRUE)*runif(nrow(incidenceData))
  
 ageData <- getBinaryTargetBaseline(
  connectionHandler = connectionHandler, 
  schema = schema,  
  analysisIds = 3
 )
 
 genderData <- getBinaryTargetBaseline(
  connectionHandler = connectionHandler, 
  schema = schema,  
  analysisIds = 1
 )

viewIncidenceRate(
  incidenceData = incidenceData,
  ageData = ageData,
  genderData = genderData
  )
#> Closing database connection
#> Closing database connection
#> Closing database connection
#> Closing database connection
#> Closing database connection
#> Closing database connection

  

Database
      	
      	Rate per 100 person years
      	
      	Proportion per 100 persons
      	N
      	Mean Follow-up Days
      	Outcomes
      	Age
      	Male %
    

Celecoxib  ( 1 ) -  - GI bleed ( wo 9999) during ( start + 0 ) - ( end + 0 )
    

Synthea
	1.369
	
	0.402
	
	1800
	1
	0
	
	27.1

Synthea
	1.369
	
	0.402
	
	1800
	1
	0
	
	22.4

Celecoxib  ( 1 ) -  - GI bleed ( wo 9999) during ( start + 0 ) - ( start + 365 )
    

Synthea
	1.371
	
	0.363
	
	1800
	300
	0
	
	27.1

Synthea
	1.371
	
	0.363
	
	1800
	300
	0
	
	22.4

Celecoxib - age 18 to 64 Age 18 to 64  ( 1001 ) -  - GI bleed ( wo 9999) during ( start + 0 ) - ( end + 0 )
    

Synthea
	1.538
	
	0.220
	
	1800
	1
	0

	NA

Celecoxib - age 18 to 64 Age 18 to 64  ( 1001 ) -  - GI bleed ( wo 9999) during ( start + 0 ) - ( start + 365 )
    

Synthea
	1.491
	
	0.324
	
	1800
	300
	0

	NA

Diclofenac  ( 2 ) -  - GI bleed ( wo 9999) during ( start + 0 ) - ( end + 0 )
    

Synthea
	1.737
	
	0.085
	
	830
	1
	0
	
	25.7

Synthea
	1.737
	
	0.085
	
	830
	1
	0
	
	19.8

Diclofenac  ( 2 ) -  - GI bleed ( wo 9999) during ( start + 0 ) - ( start + 365 )
    

Synthea
	0.031
	
	0.357
	
	830
	314
	0
	
	25.7

Synthea
	0.031
	
	0.357
	
	830
	314
	0
	
	19.8

Diclofenac - age 18 to 64 Age 18 to 64  ( 2001 ) -  - GI bleed ( wo 9999) during ( start + 0 ) - ( end + 0 )
    

Synthea
	1.297
	
	−0.314
	
	830
	1
	0

	NA

Diclofenac - age 18 to 64 Age 18 to 64  ( 2001 ) -  - GI bleed ( wo 9999) during ( start + 0 ) - ( start + 365 )
    

Synthea
	0.438
	
	−0.255
	
	830
	314
	0

	NA